+++ Git-Collab/src/App.jsx
@@ -1,100 +1,15 @@
-import React, { useState } from 'react';
 
-const Login = () => {
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
+import './App.css'
+import Login from './components/login'
 
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    // Add login logic here
-    console.log('Email:', email);
-    console.log('Password:', password);
-  };
+function App() {
+
 
   return (
-    <div style={styles.container}>
-      <div style={styles.loginBox}>
-        <h2 style={styles.title}>Sign In</h2>
-        <form onSubmit={handleSubmit}>
-          <div style={styles.inputContainer}>
-            <label style={styles.label}>Email Address</label>
-            <input
-              type="email"
-              placeholder="Enter your email"
-              value={email}
-              onChange={(e) => setEmail(e.target.value)}
-              style={styles.input}
-              required
-            />
-          </div>
-          <div style={styles.inputContainer}>
-            <label style={styles.label}>Password</label>
-            <input
-              type="password"
-              placeholder="Enter your password"
-              value={password}
-              onChange={(e) => setPassword(e.target.value)}
-              style={styles.input}
-              required
-            />
-          </div>
-          <button type="submit" style={styles.button}>
-            Sign In
-          </button>
-        </form>
-      </div>
-    </div>
-  );
-};
+    <>
+    <Login/>
+    </>
+  )
+}
 
-const styles = {
-  container: {
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
-    height: '100vh',
-    backgroundColor: '#f4f4f4',
-  },
-  loginBox: {
-    width: '350px',
-    padding: '40px',
-    backgroundColor: '#fff',
-    boxShadow: '0px 0px 10px rgba(0, 0, 0, 0.1)',
-    borderRadius: '10px',
-  },
-  title: {
-    textAlign: 'center',
-    marginBottom: '20px',
-    color: '#333',
-  },
-  inputContainer: {
-    marginBottom: '15px',
-  },
-  label: {
-    display: 'block',
-    marginBottom: '5px',
-    fontWeight: 'bold',
-    color: '#333',
-  },
-  input: {
-    width: '100%',
-    padding: '10px',
-    borderRadius: '5px',
-    border: '1px solid #ccc',
-    fontSize: '16px',
-    boxSizing: 'border-box',
-  },
-  button: {
-    width: '100%',
-    padding: '10px',
-    backgroundColor: '#007BFF',
-    color: '#fff',
-    fontSize: '16px',
-    border: 'none',
-    borderRadius: '5px',
-    cursor: 'pointer',
-    marginTop: '20px',
-  },
-};
-
-export default Login;
+export default App
